Clean Up Mess

+++ cogs/moderation/warns.py
@@ -7,6 +7,8 @@ from discord import default_permissions
 from discord.ext.commands import MissingPermissions
 from datetime import timedelta
 
+rolelist = [589435378147262464, 648546626637398046, 632674518317531137, 571032502181822506, 697002610892341298]
+
 class warns(commands.Cog):
 
     def __init__(self, client):
@@ -14,16 +16,20 @@ class warns(commands.Cog):
 
  
     @commands.command(name="warn")
-    @commands.has_permissions(manage_messages=True)
+    #@commands.has_permissions(manage_messages=True)
     async def warn(self, ctx, member: discord.Member, *, reason = None):
-        await ctx.message.delete()
-        if reason != None:
-            moderator = ctx.author
-            await self.new_warn_member(member)
-            await self.update_warns(member, reason, moderator)
-            await ctx.send(f"Warned {member.mention} for {reason}", delete_after=10)
+        user = ctx.author
+        if any(role.id in rolelist for role in user.roles):
+            await ctx.message.delete()
+            if reason != None:
+                moderator = ctx.author
+                await self.new_warn_member(member)
+                await self.update_warns(member, reason, moderator)
+                await ctx.send(f"Warned {member.mention} for {reason}", delete_after=10)
+            else:
+                await ctx.send(f"You have to specify a reason -_-", delete_after=10)
         else:
-            await ctx.send(f"You have to specify a reason -_-", delete_after=10)
+            return
 
 
     @commands.slash_command(name="warn", description="Moderator only")
@@ -39,7 +45,7 @@ class warns(commands.Cog):
     @warn.error
     async def warnerror(ctx, error):
         if isinstance(error, MissingPermissions):
-            await ctx.respond("You can't do this! You need to have moderate members permissions!", delete_after=10)
+            await ctx.respond("You can't do this! You have to have moderate members permissions!", delete_after=10)
         else:
             raise error
 
@@ -100,10 +106,11 @@ class warns(commands.Cog):
                 warns[str(member.id)][f"warn {warn_count_new}"] = f"{reason} (1st warn)"
 
                 try:
-                    await member.send(f"You were warned for the first time in the PC Creator server\nreason: {reason}")
+                    await member.send(f"You have been warned on **PC CREATOR** for `{reason}`")
+                    await the_other_logs.send(f"{member.mention} has been warned for the first time by {moderator.mention} for `{reason}`")
                 except:
-                    pass
-                await the_other_logs.send(f"{member.mention} has been warned for the first time by {moderator.mention} for **{reason}**")
+                    await the_other_logs.send(f"{member.mention} has been warned for the first time by {moderator.mention} for `{reason}`\n**DM NOT SENT**")
+                
 
 
             if warn_count_new == 2:
@@ -112,10 +119,10 @@ class warns(commands.Cog):
                 warns[str(member.id)][f"mute {mute_count_new}"] = f"{reason} (2nd warn)"
 
                 try:
-                    await member.send(f"You were warned for the second time in the PC Creator server. This leads to a 6 hour timeout\nreason: {reason}")
+                    await member.send(f"You have been warned on **PC CREATOR** for `{reason}`\nThis is your second warn, that's why you have been muted for 6 hours.")
+                    await the_other_logs.send(f"{member.mention} has been warned for the second time (mute 6h) by {moderator.mention} for `{reason}`")
                 except:
-                    pass
-                await the_other_logs.send(f"{member.mention} has been warned for the second time (mute 6 hours) by {moderator.mention} for **{reason}**")
+                    await the_other_logs.send(f"{member.mention} has been warned for the second time (mute 6h) by {moderator.mention} for `{reason}`\n**DM NOT SENT**")
 
                 duration = timedelta(hours = 6)
 
@@ -135,10 +142,10 @@ class warns(commands.Cog):
                 warns[str(member.id)][f"mute {mute_count_new}"] = f"{reason} (3rd warn)"
 
                 try:
-                    await member.send(f"You were warned for the third time in the PC Creator server. This leads to a 24 hour timeout\nreason: {reason}")
+                    await member.send(f"You have been warned on **PC CREATOR** for `{reason}`\nThis is your third warn, that's why you have been muted for 24 hours.")
+                    await the_other_logs.send(f"{member.mention} has been warned for the third time (mute 6h) by {moderator.mention} for `{reason}`")
                 except:
-                    pass
-                await the_other_logs.send(f"{member.mention} has been warned for the third time (mute 24 hours) by {moderator.mention} for **{reason}**")
+                    await the_other_logs.send(f"{member.mention} has been warned for the third time (mute 6h) by {moderator.mention} for `{reason}`\n**DM NOT SENT**")
 
                 duration = timedelta(hours = 24)
 
@@ -158,13 +165,13 @@ class warns(commands.Cog):
                 warns[str(member.id)][f"kick {kick_count_new}"] = f"{reason} (4th warn)"
 
                 try:
-                    await member.send(f"You were warned for the fourth time in the PC Creator server. This leads to a softban\nreason: {reason}")
+                    await member.send(f"You have been warned on **PC CREATOR** for `{reason}`\nThis is your fourth warn, that's why you have been soft-banned.")
+                    await the_other_logs.send(f"{member.mention} has been warned for the fourth time (softban) by {moderator.mention} for `{reason}`")
                 except:
-                    pass
-                await the_other_logs.send(f"{member.mention} has been warned for the fourth time (softban) by {moderator.mention} for **{reason}**")
+                    await the_other_logs.send(f"{member.mention} has been warned for the fourth time (softban) by {moderator.mention} for `{reason}`\n**DM NOT SENT**")
 
                 if member.guild_permissions.manage_messages:
-                    await the_other_logs.send(f"{moderator.mention} You can't ban a member with Moderator permissions.", delete_after=10)    
+                    await the_other_logs.send(f"{moderator.mention}\nYou can't ban a member with Moderator permissions.")    
 
                 await member.ban(reason=reason)
                 await member.unban(reason=reason)
@@ -181,10 +188,10 @@ class warns(commands.Cog):
                 warns[str(member.id)][f"warn {warn_count_new}"] = f"{reason} (5th warn)"
 
                 try:
-                    await member.send(f"You were warned for the fifth time in the PC Creator server. This is your **last warning**. The next warn will lead to a permanent ban.\nreason: {reason}")
+                    await member.send(f"You have been warned on **PC CREATOR** for `{reason}`\nThis is your last warning, please behave or you may get banned.")
+                    await the_other_logs.send(f"{member.mention} has been warned for the fifth time (last warn) by {moderator.mention} for `{reason}`")
                 except:
-                    pass
-                await the_other_logs.send(f"{member.mention} has been warned for the fifth time (last warning) by {moderator.mention} for **{reason}**")
+                    await the_other_logs.send(f"{member.mention} has been warned for the fifth time (last warn) by {moderator.mention} for `{reason}`\n**DM NOT SENT**")
 
             if warn_count_new == 6:
 
@@ -192,13 +199,13 @@ class warns(commands.Cog):
                 warns[str(member.id)][f"ban {ban_count_new}"] = f"{reason} (6th warn)"
 
                 try:
-                    await member.send(f"You were warned for the sixth time in the PC Creator server. This leads to a **permanent ban**. If you think there is a mistake DM <@695229647021015040>\nreason: {reason}")
+                    await member.send(f"You have been banned from **PC CREATOR** for `{reason}`")
+                    await the_other_logs.send(f"{member.mention} has been warned for the sixth time (ban) by {moderator.mention} for `{reason}`")
                 except:
-                    pass
-                await the_other_logs.send(f"{member.mention} has been warned for the sixth time (permanent ban) by {moderator.mention} for **{reason}**")
+                    await the_other_logs.send(f"{member.mention} has been warned for the sixth time (ban) by {moderator.mention} for `{reason}`\n**DM NOT SENT**")
 
                 if member.guild_permissions.manage_messages:
-                    await the_other_logs.send(f"{moderator.mention} You can't ban a member with Moderator permissions.", delete_after=10)    
+                    await the_other_logs.send(f"{moderator.mention} You can't ban a member with Moderator permissions.")    
 
                 await member.ban(reason=reason)
 
@@ -207,6 +214,13 @@ class warns(commands.Cog):
                 embed.add_field(name="Moderator", value=f"{moderator.mention}")
                 embed.add_field(name="Reason:", value=reason, inline=False)
                 await modlogs.send(embed=embed)
+       
+            else:
+                try:
+                    await member.send(f"You have been banned from **PC CREATOR** for `{reason}`")
+                    await the_other_logs.send(f"{member.mention} has been warned for the {warn_count_new}th time by {moderator.mention} for `{reason}`")
+                except:
+                    await the_other_logs.send(f"{member.mention} has been warned for the {warn_count_new}th time by {moderator.mention} for `{reason}`\n**DM NOT SENT**")
 
         with open("json_files/warns.json", "w") as f:
             json.dump(warns,f)
